refactor common component
