+++ frontend/src/components/auth/reset-password/ForgotPassword.tsx
@@ -1,22 +1,16 @@
 import React, { useState } from "react";
 import { useMutation, useQuery } from "@apollo/client";
-import {
-  Button,
-  FormControl,
-  FormLabel,
-  Input,
-  Text,
-  useToast,
-} from "@chakra-ui/react";
+import { FormControl, FormLabel, Input } from "@chakra-ui/react";
 
 import { RESET_PASSWORD } from "../../../APIClients/mutations/AuthMutations";
 import { GET_USERS_BY_ROLE } from "../../../APIClients/queries/UserQueries";
 import { UserResponse } from "../../../APIClients/types/UserClientTypes";
-import { ArrowBackOutlineIcon } from "../../../assets/icons";
 import {
   ADMIN_SIGNUP_IMAGE,
   TEACHER_SIGNUP_IMAGE,
 } from "../../../assets/images";
+import LoadingState from "../../common/LoadingState";
+import ResendEmail from "../../common/ResendEmail";
 import AuthWrapper from "../AuthWrapper";
 import NavigationButtons from "../teacher-signup/NavigationButtons";
 
@@ -32,11 +26,9 @@ const ForgotPassword = ({
   const [emailNotFoundError, setEmailNotFoundError] = useState(false);
   const [step, setStep] = useState(1);
 
-  const toast = useToast();
   const { data } = useQuery(GET_USERS_BY_ROLE, {
     variables: { role: isAdmin ? "Admin" : "Teacher" },
   });
-
   const [resetPassword] = useMutation<{ resetPassword: boolean }>(
     RESET_PASSWORD,
     {
@@ -58,17 +50,6 @@ const ForgotPassword = ({
     }
 
     await resetPassword({ variables: { email } });
-  };
-
-  const onResendClick = async () => {
-    await onResetPasswordClick();
-    toast({
-      title: "Resent email!",
-      description: "Check your email to reset your password.",
-      status: "success",
-      duration: 9000,
-      isClosable: true,
-    });
   };
 
   const title = step === 1 ? "Forgot Password?" : "Check your email";
@@ -102,28 +83,7 @@ const ForgotPassword = ({
           />
         </>
       )}
-      {step === 2 && (
-        <>
-          <Text textAlign="center" textStyle="subtitle2">
-            Didn’t receive the email?{" "}
-            <Button
-              color="blue.300"
-              display="contents"
-              onClick={onResendClick}
-              style={{ font: "inherit" }}
-            >
-              Click to resend
-            </Button>
-          </Text>
-          <Button
-            leftIcon={<ArrowBackOutlineIcon />}
-            onClick={() => window.location.reload()}
-            variant="tertiary"
-          >
-            Back to login page
-          </Button>
-        </>
-      )}
+      {step === 2 && <ResendEmail resendFunction={onResetPasswordClick} />}
     </>
   );
